+++ edge/extension.ts
@@ -11,6 +11,7 @@ import commitAmend from './commitAmend'
 import commitEmpty from './commitEmpty'
 import urgent, { urgentRestore } from './urgent'
 import stash, { stashPopLatest, stashPop, stashClear, updateStashCountBar } from './stash'
+import squash from './squash'
 import master from './master'
 import branch from './branch'
 import checkout from './checkout'
@@ -55,6 +56,8 @@ export async function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(vscode.commands.registerCommand('gitGrace.stashPop', Queue.put(stashPop)))
 
     context.subscriptions.push(vscode.commands.registerCommand('gitGrace.stashClear', Queue.put(stashClear)))
+
+    context.subscriptions.push(vscode.commands.registerCommand('gitGrace.squash', squash))
 
     context.subscriptions.push(vscode.commands.registerCommand('gitGrace.master', Queue.put(master, [fetch])))